Move test fixture to conftest
