+++ forest/benchmarking/tests/test_process_tomography.py
@@ -1,10 +1,6 @@
-import networkx as nx
 import numpy as np
 import pytest
-from requests.exceptions import RequestException
-from rpcq.messages import PyQuilExecutableResponse
 
-from forest.benchmarking.compilation import basic_compile
 from forest.benchmarking.operator_tools.random_operators import haar_rand_unitary
 from forest.benchmarking.operator_tools.superoperator_transformations import kraus2choi
 from forest.benchmarking.tomography import generate_process_tomography_experiment, \
@@ -14,39 +10,8 @@ from forest.benchmarking.observable_estimation import estimate_observables, Expe
     _one_q_state_prep
 from pyquil import Program
 from pyquil import gate_matrices as mat
-from pyquil.api import QVM
 from pyquil.gates import CNOT, X
 from pyquil.numpy_simulator import NumpyWavefunctionSimulator
-
-
-@pytest.fixture
-def test_qc():
-    from pyquil.api import ForestConnection, QuantumComputer
-    from pyquil.api._compiler import _extract_attribute_dictionary_from_program
-    from pyquil.api._qac import AbstractCompiler
-    from pyquil.device import NxDevice
-    from pyquil.gates import I
-
-    class BasicQVMCompiler(AbstractCompiler):
-
-        def quil_to_native_quil(self, program: Program, protoquil=None):
-            return basic_compile(program)
-
-        def native_quil_to_executable(self, nq_program: Program):
-            return PyQuilExecutableResponse(
-                program=nq_program.out(),
-                attributes=_extract_attribute_dictionary_from_program(nq_program))
-    try:
-        qc = QuantumComputer(
-            name='testing-qc',
-            qam=QVM(connection=ForestConnection(), random_seed=52),
-            device=NxDevice(nx.complete_graph(2)),
-            compiler=BasicQVMCompiler(),
-        )
-        qc.run_and_measure(Program(I(0)), trials=1)
-        return qc
-    except (RequestException, TimeoutError) as e:
-        return pytest.skip("This test requires a running local QVM: {}".format(e))
 
 
 def wfn_estimate_observables(n_qubits, tomo_expt: ObservablesExperiment):
